add condition

+++ files/routes/admin.py
@@ -1884,7 +1884,8 @@ def delete_media_post(v):
 		)
 	g.db.add(ma)
 
-	purge_files_in_cloudflare_cache(url)
+	if not url.startswith('https://videos'):
+		purge_files_in_cloudflare_cache(url)
 
 	if SITE == 'watchpeopledie.tv' and url.startswith(SITE_FULL_VIDEOS):
 		filename = url.split(SITE_FULL_VIDEOS)[1]
